requirements.txt
groq==0.31.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
